pyproject.toml
[tool.poetry]
name = "infrastructure-optimizer"
version = "0.1.0"
description = "Solution d'optimisation d'infrastructure technique pour Devoteam"
authors = ["Candidate Devoteam"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10"
django = ">=5.2.6,<6.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
python-decouple = ">=3.8,<4.0"
openai = ">=1.107.1,<2.0.0"
requests = ">=2.32.5,<3.0.0"
numpy = "<2.0"
pandas = "<3.0"
scikit-learn = "<2.0"
drf-yasg = ">=1.21.10,<2.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
